Move config files to a `config/` folder
