+++ src/config/keyboard-shortcut-config.js
